# 前日(JST)の自動適用（html/all）・自動化の高速化・README更新