## 概要\n- 日付未指定時に前日(JST)を自動適用（html/all）\n- 自動化の待機時間を短縮し応答性を改善\n- READMEに挙動を追記（使用例/仕様/FAQ/更新履歴）\n\n## 変更内容\n- feat(main): 日付未指定時は前日(JST)を自動適用（html/all）\n  - --no-date 未指定かつ日付引数なしの場合、JSTの前日を YYMMDD で自動設定\n  - --no-date の挙動は変更なし\n  - --verbose 時にINFO出力\n- perf(create_html): 待機時間短縮（0.5s→0.2s, 1s→0.5s, 2s→0.5s など）\n- docs(readme): 日付省略時の前日(JST)自動適用を明記し、例・仕様・FAQ・更新履歴を更新\n\n## 目的\n- 日次運用の手間削減と体感速度向上\n\n## 影響範囲\n- 仕様追加（デフォルト日付）・性能改善・ドキュメント\n- 既存の --no-date 挙動は変更なし\n\n## 動作確認\n- 2025-09-28 実行時に 2025-09-27 の since/until が自動設定\n- 自動化フローのタイミング短縮で正常動作を確認\n\n## 備考\n- data/config/positions.json はコミットから除外\n